imx547 : Used devm_* resource management functions instead of gpiod_*
With current use of non-devm functions the user is responsible for freeing
the resources and since this is not handled currently the driver does not probe
the sensor on the second load, since resources are not freed.
Here is a dmesg snippet:
On first load
ubuntu@kria:~$ sudo dmesg | grep imx547
[   72.621037] imx547: module is from the staging directory, the quality is unknown, you have been warned.
[   72.628992] imx547 7-001a: imx547 : imx547 probe success !
On second load:
ubuntu@kria:~$ sudo dmesg | grep imx547
[  312.319586] imx547 7-001a: GT TRX Reset GPIO not setup in DT
[  312.325487] imx547: probe of 7-001a failed with error -16

Since freeing of the GPIO resource is not explicitly handled in current implementation
the driver fails to probe the sensor and fails to add it to media subsystem. In this case,
developer needs to add explicit cleanup code.

New implementation of using device managed ie devm_* functions ensures that resource is
automatically removed when driver is unloaded.

Signed-off-by: Sagar Karmarkar <sagar.karmarkar@amd.com>